+++ game/bishop.py
@@ -1,9 +1,7 @@
 from game.piece import Piece
 
 class Bishop(Piece):
-    def __init__(self, color):
-        super().__init__(color)
-        self.__type__ = "BISHOP"
+    __type__ = "BISHOP"
 
     def permited_move(self, from_row, from_col, to_row, to_col, board):
         permited_move_diag = self.permited_move_diagonal(from_row, from_col, to_row, to_col, board)