+++ game/main.py
@@ -17,9 +17,9 @@ class Cli():
                 from_row = int(input("From row: "))
                 from_col = int(input("From col: "))
               
-                self.chess.error_out_of_range(from_row, from_col)
+                self.chess.verify_out_of_range(from_row, from_col)
 
-                print("The piece you have chosen is: ", chess.__board__.get_piece(from_row, from_col))
+                print("The piece you have chosen is: ", chess.__board__.get_piece_for_show(from_row, from_col))
             
                 # Verificamos el color de la pieza usando la nueva función
                 if self.verify_color(chess, from_row, from_col):
@@ -52,7 +52,6 @@ class Cli():
         while a == "y":
             self.chess.__board__.show_board() 
             try:
-
                 from_row, from_col = self.verify_move(self.chess)
 
                 to_row, to_col = self.validate_range_to()
@@ -67,17 +66,14 @@ class Cli():
             
                 print(self.chess.show_eaten_pieces())
 
-                if self.chess.verify_winner() != False:
+                if self.chess.verify_winner() is not False:
                     print(self.chess.verify_winner())
                     a = "n"
                     break
-
                 a = input("Do you want to continue? (y/n): ")
                 if a == "y":
                     self.chess.change_turn()
                     print("Es turno de: ", self.chess.__turn__)
-                else:
-                    raise GameEnded("Game ended")
 
             except (NotPieceToMove, NotPermitedMove, InvalidPosition, NotPieceToReplace) as e:
                 print("Error:", e)
@@ -86,18 +82,17 @@ class Cli():
             except Exception as e:
                 print("error", e)
                 return "error"
-            except GameEnded as e:
-                print("Game ended. Noone wins")
-            # return "end"
+            
+        print("Game ended")
+
 
     def validate_range_to(self):
-
         while True:
             try:
                 to_row = int(input("To row: "))
                 to_col = int(input("To col: "))
 
-                self.chess.error_out_of_range(to_row, to_col)
+                self.chess.verify_out_of_range(to_row, to_col)
 
                 return to_row, to_col